added synthesizer test files for better test organization
