+++ src/synthesizer/csharp/mod.rs
@@ -782,113 +782,4 @@ impl OutputInstruction {
 }
 
 #[cfg(test)]
-mod tests {
-    use std::borrow::Cow;
-
-    use crate::{
-        cdk::Schema,
-        code::CodeBuffer,
-        ir::{conditions::ConditionIr, importer::ImportInstruction, resources::ResourceIr},
-        primitives::WrapperF64,
-    };
-
-    use super::CsharpEmitter;
-
-    #[test]
-    fn test_fn_split() {
-        let output = CodeBuffer::default();
-        let schema = Cow::Borrowed(Schema::builtin());
-        let resource_ir = ResourceIr::Split(
-            "-".into(),
-            Box::new(ResourceIr::String("My-EC2-Instance".into())),
-        );
-        let result = resource_ir.emit_csharp(&output, &schema);
-        assert_eq!((), result.unwrap());
-    }
-
-    #[test]
-    fn test_fn_split_other() {
-        let output = CodeBuffer::default();
-        let schema = Cow::Borrowed(Schema::builtin());
-        let resource_ir = ResourceIr::Split(
-            "-".into(),
-            Box::new(ResourceIr::Join(
-                ",".to_string(),
-                vec![
-                    ResourceIr::String("a".into()),
-                    ResourceIr::String("b".into()),
-                    ResourceIr::String("c".into()),
-                ],
-            )),
-        );
-        let result = resource_ir.emit_csharp(&output, &schema);
-        assert_eq!((), result.unwrap());
-    }
-
-    #[test]
-    fn test_condition_ir_map() {
-        let output = CodeBuffer::default();
-        let schema = Cow::Borrowed(Schema::builtin());
-        let condition_ir = ConditionIr::Map(
-            "ConditionIrMap".into(),
-            Box::new(ConditionIr::Str("FirstLevelKey".into())),
-            Box::new(ConditionIr::Str("SecondLevelKey".into())),
-        );
-        let result = condition_ir.emit_csharp(&output, &schema);
-        assert_eq!((), result.unwrap());
-    }
-
-    #[test]
-    fn test_condition_ir_split() {
-        let output = CodeBuffer::default();
-        let schema = Cow::Borrowed(Schema::builtin());
-        let condition_ir = ConditionIr::Split(
-            "-".into(),
-            Box::new(ConditionIr::Str("string-to-split".into())),
-        );
-        let result = condition_ir.emit_csharp(&output, &schema);
-        assert_eq!((), result.unwrap());
-    }
-
-    #[test]
-    fn test_alexa_org() {
-        let import_instruction = ImportInstruction {
-            organization: "Alexa".into(),
-            service: Some("Ask".into()),
-        };
-        let result = import_instruction.to_csharp();
-        assert_eq!("using Amazon.CDK.Alexa.Ask;", result.unwrap());
-    }
-
-    #[test]
-    fn test_resource_ir_double() {
-        let output = CodeBuffer::default();
-        let schema = Cow::Borrowed(Schema::builtin());
-        let resource_ir = ResourceIr::Double(WrapperF64::new(2.0));
-        let result = resource_ir.emit_csharp(&output, &schema);
-        assert_eq!((), result.unwrap());
-    }
-
-    #[test]
-    fn test_resource_ir_select() {
-        let output = CodeBuffer::default();
-        let schema = Cow::Borrowed(Schema::builtin());
-        let resource_ir =
-            ResourceIr::Select(1, Box::new(ResourceIr::String("Not an array".into())));
-        let result = resource_ir.emit_csharp(&output, &schema);
-        assert_eq!((), result.unwrap());
-    }
-
-    #[test]
-    fn test_resource_ir_cidr() {
-        let output = CodeBuffer::default();
-        let schema = Cow::Borrowed(Schema::builtin());
-        let resource_ir = ResourceIr::Cidr(
-            Box::new(ResourceIr::String("0.0.0.0".into())),
-            Box::new(ResourceIr::String("16".into())),
-            Box::new(ResourceIr::String("255.255.255.0".into())),
-        );
-        let result = resource_ir.emit_csharp(&output, &schema);
-        assert_eq!((), result.unwrap());
-    }
-}
+mod tests;
