+++ src/synthesizer/typescript/tests.rs
@@ -10,3 +10,15 @@ fn pretty_name_fixes() {
     assert_eq!("targetAZs", pretty_name("TargetAZs"));
     assert_eq!("diskSizeMBs", pretty_name("DiskSizeMBs"));
 }
+
+#[test]
+fn test_invalid_organization() {
+  let bad_org = "NotAws";
+  let import_instruction = ImportInstruction {
+    organization: bad_org.to_string(),
+    service: Option::None,
+  };
+  let result = import_instruction.to_typescript().unwrap_err();
+  let expected = format!("Expected organization to be AWS or Alexa. Found {bad_org}");
+  assert_eq!(expected, result.to_string());
+}
