+++ tests/tests.rs
@@ -1,6 +1,5 @@
-use cdk_from_cfn::cdk::{Primitive, Schema, TypeReference};
+use cdk_from_cfn::cdk::Schema;
 use cdk_from_cfn::ir::importer::ImportInstruction;
-use cdk_from_cfn::ir::resources::{ResourceInstruction, ResourceIr, ResourceType};
 use cdk_from_cfn::ir::CloudformationProgramIr;
 use cdk_from_cfn::parser::condition::{ConditionFunction, ConditionValue};
 use cdk_from_cfn::parser::lookup_table::{MappingInnerValue, MappingTable};
@@ -11,7 +10,6 @@ use cdk_from_cfn::primitives::WrapperF64;
 use cdk_from_cfn::{synthesizer, CloudformationParseTree};
 use indexmap::IndexMap;
 use std::borrow::Cow;
-use std::collections::BTreeSet;
 use std::vec;
 
 mod json;
@@ -834,166 +832,5 @@ fn test_resource_translation_error() {
         Named(\"AWS::Serverless::Function.HttpApiEvent\"), \
         Named(\"AWS::Serverless::Function.CognitoEvent\")\
         ]))) is not implemented for ResourceValue::Object";
-    assert_eq!(expected, result.to_string());
-}
-
-#[test]
-fn test_invalid_organization_for_typescript() {
-    let bad_org = "NotAWS";
-    let ir = CloudformationProgramIr {
-        imports: vec![ImportInstruction {
-            organization: "NotAWS".to_string(),
-            service: None,
-        }],
-        ..Default::default()
-    };
-    let synthesizer = Box::new(crate::synthesizer::Typescript {});
-    let mut output = Vec::new();
-    let result = ir
-        .synthesize(synthesizer.as_ref(), &mut output, "Stack")
-        .unwrap_err();
-    let expected = format!("Expected organization to be AWS or Alexa. Found {bad_org}");
-    assert_eq!(expected, result.to_string());
-}
-
-#[test]
-fn test_invalid_organization_for_python() {
-    let bad_org = "NotAWS";
-    let ir = CloudformationProgramIr {
-        imports: vec![ImportInstruction {
-            organization: "NotAWS".to_string(),
-            service: None,
-        }],
-        ..Default::default()
-    };
-    let synthesizer = Box::new(crate::synthesizer::Python {});
-    let mut output = Vec::new();
-    let result = ir
-        .synthesize(synthesizer.as_ref(), &mut output, "Stack")
-        .unwrap_err();
-    let expected = format!("Expected organization to be AWS or Alexa. Found {bad_org}");
-    assert_eq!(expected, result.to_string());
-}
-
-#[test]
-fn test_invalid_organization_for_csharp() {
-    let bad_org = "NotAWS";
-    let ir = CloudformationProgramIr {
-        imports: vec![ImportInstruction {
-            organization: "NotAWS".to_string(),
-            service: None,
-        }],
-        ..Default::default()
-    };
-    let synthesizer = Box::<crate::synthesizer::CSharp>::default();
-    let mut output = Vec::new();
-    let result = ir
-        .synthesize(synthesizer.as_ref(), &mut output, "Stack")
-        .unwrap_err();
-    let expected = format!("Expected organization to be AWS or Alexa. Found {bad_org}");
-    assert_eq!(expected, result.to_string());
-}
-
-#[test]
-fn test_invalid_organization_for_java() {
-    let bad_org = "NotAWS";
-    let ir = CloudformationProgramIr {
-        imports: vec![ImportInstruction {
-            organization: "NotAWS".to_string(),
-            service: None,
-        }],
-        ..Default::default()
-    };
-    let synthesizer = Box::<crate::synthesizer::Java>::default();
-    let mut output = Vec::new();
-    let result = ir
-        .synthesize(synthesizer.as_ref(), &mut output, "Stack")
-        .unwrap_err();
-    let expected = format!("Expected organization to be AWS or Alexa. Found {bad_org}");
-    assert_eq!(expected, result.to_string());
-}
-
-#[test]
-fn test_invalid_organization_for_golang() {
-    let bad_org = "NotAWS";
-    let ir = CloudformationProgramIr {
-        imports: vec![ImportInstruction {
-            organization: "NotAWS".to_string(),
-            service: None,
-        }],
-        ..Default::default()
-    };
-    let synthesizer = Box::<crate::synthesizer::Golang>::default();
-    let mut output = Vec::new();
-    let result = ir
-        .synthesize(synthesizer.as_ref(), &mut output, "Stack")
-        .unwrap_err();
-    let expected = format!("Expected organization to be AWS or Alexa. Found {bad_org}");
-    assert_eq!(expected, result.to_string());
-}
-
-#[test]
-fn test_invalid_resource_object_structure() {
-    let resource_ir = ResourceIr::Object(
-        TypeReference::Union(cdk_from_cfn::cdk::TypeUnion::Static(&[])),
-        IndexMap::default(),
-    );
-    let properties = IndexMap::from([("BadProperty".into(), resource_ir)]);
-    let ir = CloudformationProgramIr {
-        resources: vec![ResourceInstruction {
-            name: "InvalidResource".to_string(),
-            condition: Option::None,
-            metadata: Option::None,
-            update_policy: Option::None,
-            deletion_policy: Option::None,
-            dependencies: vec![],
-            resource_type: ResourceType::AWS {
-                service: "Dynamo".to_string(),
-                type_name: "GlobalTable".to_string(),
-            },
-            properties,
-            references: BTreeSet::new(),
-        }],
-        ..Default::default()
-    };
-    let synthesizer = Box::<crate::synthesizer::CSharp>::default();
-    let mut output = Vec::new();
-    let result = ir
-        .synthesize(synthesizer.as_ref(), &mut output, "Stack")
-        .unwrap_err();
-    let expected = "Type reference Union(\n    Static(\n        [],\n    ),\n) not implemented for ResourceIr::Object";
-    assert_eq!(expected, result.to_string());
-}
-
-#[test]
-fn test_invalid_resource_object_primitive() {
-    let resource_ir = ResourceIr::Object(
-        TypeReference::Primitive(Primitive::String),
-        IndexMap::default(),
-    );
-    let properties = IndexMap::from([("BadProperty".into(), resource_ir)]);
-    let ir = CloudformationProgramIr {
-        resources: vec![ResourceInstruction {
-            name: "InvalidResource".to_string(),
-            condition: Option::None,
-            metadata: Option::None,
-            update_policy: Option::None,
-            deletion_policy: Option::None,
-            dependencies: vec![],
-            resource_type: ResourceType::AWS {
-                service: "Dynamo".to_string(),
-                type_name: "GlobalTable".to_string(),
-            },
-            properties,
-            references: BTreeSet::new(),
-        }],
-        ..Default::default()
-    };
-    let synthesizer = Box::<crate::synthesizer::CSharp>::default();
-    let mut output = Vec::new();
-    let result = ir
-        .synthesize(synthesizer.as_ref(), &mut output, "Stack")
-        .unwrap_err();
-    let expected = "Cannot emit ResourceIr::Object with non-json simple structure (String)";
     assert_eq!(expected, result.to_string());
 }
